Buffer current and append (ary_of_strings)
it's possible an array of strings isn't the only thing you would want to
pass here, maybe you would want to copy lines from one buffer to another
for instance and then this api would need to conditionally support it or
it could become awkward. hard to say. perhaps an append_lines version of
the method or something would make sense